ENH: mail collectors: common lib
